+++ src/hyperspell/types/memory.py
@@ -1,8 +1,10 @@
 # File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.
 
-from typing import TYPE_CHECKING, List, Optional
+from typing import TYPE_CHECKING, Dict, List, Optional
 from datetime import datetime
 from typing_extensions import Literal
+
+from pydantic import Field as FieldInfo
 
 from .._models import BaseModel
 
@@ -18,6 +20,8 @@ class MetadataEvent(BaseModel):
 
 
 class Metadata(BaseModel):
+    created_at: Optional[datetime] = None
+
     events: Optional[List[MetadataEvent]] = None
 
     indexed_at: Optional[datetime] = None
@@ -26,6 +30,9 @@ class Metadata(BaseModel):
 
     status: Optional[Literal["pending", "processing", "completed", "failed"]] = None
 
+    url: Optional[str] = None
+
+    __pydantic_extra__: Dict[str, object] = FieldInfo(init=False)  # pyright: ignore[reportIncompatibleVariableOverride]
     if TYPE_CHECKING:
         # Stub to indicate that arbitrary properties are accepted.
         # To access properties that are not valid identifiers you can use `getattr`, e.g.