+++ src/hyperspell/types/memory_search_params.py
@@ -10,15 +10,11 @@ from .._utils import PropertyInfo
 
 __all__ = [
     "MemorySearchParams",
-    "Filter",
-    "FilterGoogleCalendar",
-    "FilterGoogleMail",
-    "FilterNotion",
-    "FilterReddit",
-    "FilterSlack",
-    "FilterWebCrawler",
     "Options",
+    "OptionsBox",
+    "OptionsCollections",
     "OptionsGoogleCalendar",
+    "OptionsGoogleDrive",
     "OptionsGoogleMail",
     "OptionsNotion",
     "OptionsReddit",
@@ -33,12 +29,6 @@ class MemorySearchParams(TypedDict, total=False):
 
     answer: bool
     """If true, the query will be answered along with matching source documents."""
-
-    filter: Optional[Filter]
-    """DEPRECATED: Use options instead.
-
-    This field will be removed in a future version.
-    """
 
     max_results: int
     """Maximum number of results to return."""
@@ -99,102 +89,45 @@ class MemorySearchParams(TypedDict, total=False):
     """Only query documents from these sources."""
 
 
-class FilterGoogleCalendar(TypedDict, total=False):
-    calendar_id: Optional[str]
-    """The ID of the calendar to search.
-
-    If not provided, it will use the ID of the default calendar. You can get the
-    list of calendars with the `/integrations/google_calendar/list` endpoint.
-    """
-
-
-class FilterGoogleMail(TypedDict, total=False):
-    label_ids: List[str]
-    """List of label IDs to filter messages (e.g., ['INBOX', 'SENT', 'DRAFT']).
-
-    Multiple labels are combined with OR logic - messages matching ANY specified
-    label will be returned. If empty, no label filtering is applied (searches all
-    accessible messages).
-    """
-
-
-class FilterNotion(TypedDict, total=False):
-    notion_page_ids: List[str]
-    """List of Notion page IDs to search.
-
-    If not provided, all pages in the workspace will be searched.
-    """
-
-
-class FilterReddit(TypedDict, total=False):
-    period: Literal["hour", "day", "week", "month", "year", "all"]
-    """The time period to search. Defaults to 'month'."""
-
-    sort: Literal["relevance", "new", "hot", "top", "comments"]
-    """The sort order of the posts. Defaults to 'relevance'."""
-
-    subreddit: Optional[str]
-    """The subreddit to search.
-
-    If not provided, the query will be searched for in all subreddits.
-    """
-
-
-class FilterSlack(TypedDict, total=False):
-    channels: List[str]
-    """List of Slack channels to search.
-
-    If not provided, all channels in the workspace will be searched.
-    """
-
-
-class FilterWebCrawler(TypedDict, total=False):
-    max_depth: int
-    """Maximum depth to crawl from the starting URL"""
-
-    url: Union[str, object]
-    """The URL to crawl"""
-
-
-class Filter(TypedDict, total=False):
+class OptionsBox(TypedDict, total=False):
     after: Annotated[Union[str, datetime, None], PropertyInfo(format="iso8601")]
     """Only query documents created on or after this date."""
-
-    answer_model: Literal["llama-3.1", "gemma2", "qwen-qwq", "mistral-saba", "llama-4-scout", "deepseek-r1"]
-    """Model to use for answer generation when answer=True"""
 
     before: Annotated[Union[str, datetime, None], PropertyInfo(format="iso8601")]
     """Only query documents created before this date."""
 
-    box: object
-    """Search options for Box"""
+    weight: float
+    """Weight of results from this source.
 
-    collections: object
-    """Search options for vault"""
+    A weight greater than 1.0 means more results from this source will be returned,
+    a weight less than 1.0 means fewer results will be returned. This will only
+    affect results if multiple sources are queried at the same time.
+    """
 
-    google_calendar: FilterGoogleCalendar
-    """Search options for Google Calendar"""
 
-    google_drive: object
-    """Search options for Google Drive"""
+class OptionsCollections(TypedDict, total=False):
+    after: Annotated[Union[str, datetime, None], PropertyInfo(format="iso8601")]
+    """Only query documents created on or after this date."""
 
-    google_mail: FilterGoogleMail
-    """Search options for Gmail"""
+    before: Annotated[Union[str, datetime, None], PropertyInfo(format="iso8601")]
+    """Only query documents created before this date."""
 
-    notion: FilterNotion
-    """Search options for Notion"""
+    weight: float
+    """Weight of results from this source.
 
-    reddit: FilterReddit
-    """Search options for Reddit"""
-
-    slack: FilterSlack
-    """Search options for Slack"""
-
-    web_crawler: FilterWebCrawler
-    """Search options for Web Crawler"""
+    A weight greater than 1.0 means more results from this source will be returned,
+    a weight less than 1.0 means fewer results will be returned. This will only
+    affect results if multiple sources are queried at the same time.
+    """
 
 
 class OptionsGoogleCalendar(TypedDict, total=False):
+    after: Annotated[Union[str, datetime, None], PropertyInfo(format="iso8601")]
+    """Only query documents created on or after this date."""
+
+    before: Annotated[Union[str, datetime, None], PropertyInfo(format="iso8601")]
+    """Only query documents created before this date."""
+
     calendar_id: Optional[str]
     """The ID of the calendar to search.
 
@@ -202,8 +135,38 @@ class OptionsGoogleCalendar(TypedDict, total=False):
     list of calendars with the `/integrations/google_calendar/list` endpoint.
     """
 
+    weight: float
+    """Weight of results from this source.
+
+    A weight greater than 1.0 means more results from this source will be returned,
+    a weight less than 1.0 means fewer results will be returned. This will only
+    affect results if multiple sources are queried at the same time.
+    """
+
+
+class OptionsGoogleDrive(TypedDict, total=False):
+    after: Annotated[Union[str, datetime, None], PropertyInfo(format="iso8601")]
+    """Only query documents created on or after this date."""
+
+    before: Annotated[Union[str, datetime, None], PropertyInfo(format="iso8601")]
+    """Only query documents created before this date."""
+
+    weight: float
+    """Weight of results from this source.
+
+    A weight greater than 1.0 means more results from this source will be returned,
+    a weight less than 1.0 means fewer results will be returned. This will only
+    affect results if multiple sources are queried at the same time.
+    """
+
 
 class OptionsGoogleMail(TypedDict, total=False):
+    after: Annotated[Union[str, datetime, None], PropertyInfo(format="iso8601")]
+    """Only query documents created on or after this date."""
+
+    before: Annotated[Union[str, datetime, None], PropertyInfo(format="iso8601")]
+    """Only query documents created before this date."""
+
     label_ids: List[str]
     """List of label IDs to filter messages (e.g., ['INBOX', 'SENT', 'DRAFT']).
 
@@ -212,16 +175,44 @@ class OptionsGoogleMail(TypedDict, total=False):
     accessible messages).
     """
 
+    weight: float
+    """Weight of results from this source.
+
+    A weight greater than 1.0 means more results from this source will be returned,
+    a weight less than 1.0 means fewer results will be returned. This will only
+    affect results if multiple sources are queried at the same time.
+    """
+
 
 class OptionsNotion(TypedDict, total=False):
+    after: Annotated[Union[str, datetime, None], PropertyInfo(format="iso8601")]
+    """Only query documents created on or after this date."""
+
+    before: Annotated[Union[str, datetime, None], PropertyInfo(format="iso8601")]
+    """Only query documents created before this date."""
+
     notion_page_ids: List[str]
     """List of Notion page IDs to search.
 
     If not provided, all pages in the workspace will be searched.
     """
 
+    weight: float
+    """Weight of results from this source.
+
+    A weight greater than 1.0 means more results from this source will be returned,
+    a weight less than 1.0 means fewer results will be returned. This will only
+    affect results if multiple sources are queried at the same time.
+    """
+
 
 class OptionsReddit(TypedDict, total=False):
+    after: Annotated[Union[str, datetime, None], PropertyInfo(format="iso8601")]
+    """Only query documents created on or after this date."""
+
+    before: Annotated[Union[str, datetime, None], PropertyInfo(format="iso8601")]
+    """Only query documents created before this date."""
+
     period: Literal["hour", "day", "week", "month", "year", "all"]
     """The time period to search. Defaults to 'month'."""
 
@@ -234,21 +225,57 @@ class OptionsReddit(TypedDict, total=False):
     If not provided, the query will be searched for in all subreddits.
     """
 
+    weight: float
+    """Weight of results from this source.
+
+    A weight greater than 1.0 means more results from this source will be returned,
+    a weight less than 1.0 means fewer results will be returned. This will only
+    affect results if multiple sources are queried at the same time.
+    """
+
 
 class OptionsSlack(TypedDict, total=False):
+    after: Annotated[Union[str, datetime, None], PropertyInfo(format="iso8601")]
+    """Only query documents created on or after this date."""
+
+    before: Annotated[Union[str, datetime, None], PropertyInfo(format="iso8601")]
+    """Only query documents created before this date."""
+
     channels: List[str]
     """List of Slack channels to search.
 
     If not provided, all channels in the workspace will be searched.
     """
 
+    weight: float
+    """Weight of results from this source.
+
+    A weight greater than 1.0 means more results from this source will be returned,
+    a weight less than 1.0 means fewer results will be returned. This will only
+    affect results if multiple sources are queried at the same time.
+    """
+
 
 class OptionsWebCrawler(TypedDict, total=False):
+    after: Annotated[Union[str, datetime, None], PropertyInfo(format="iso8601")]
+    """Only query documents created on or after this date."""
+
+    before: Annotated[Union[str, datetime, None], PropertyInfo(format="iso8601")]
+    """Only query documents created before this date."""
+
     max_depth: int
     """Maximum depth to crawl from the starting URL"""
 
     url: Union[str, object]
     """The URL to crawl"""
+
+    weight: float
+    """Weight of results from this source.
+
+    A weight greater than 1.0 means more results from this source will be returned,
+    a weight less than 1.0 means fewer results will be returned. This will only
+    affect results if multiple sources are queried at the same time.
+    """
 
 
 class Options(TypedDict, total=False):
@@ -261,20 +288,23 @@ class Options(TypedDict, total=False):
     before: Annotated[Union[str, datetime, None], PropertyInfo(format="iso8601")]
     """Only query documents created before this date."""
 
-    box: object
+    box: OptionsBox
     """Search options for Box"""
 
-    collections: object
+    collections: OptionsCollections
     """Search options for vault"""
 
     google_calendar: OptionsGoogleCalendar
     """Search options for Google Calendar"""
 
-    google_drive: object
+    google_drive: OptionsGoogleDrive
     """Search options for Google Drive"""
 
     google_mail: OptionsGoogleMail
     """Search options for Gmail"""
+
+    max_results: int
+    """Maximum number of results to return."""
 
     notion: OptionsNotion
     """Search options for Notion"""